pyproject.toml
[tool.poetry]
name = "postmates"
version = "0.1.0"
description = ""
authors = ["jam <nevvjann@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
djangorestframework = "^3.11.0"
Pillow = "^7.1.2"
django-filter = "^2.3.0"
gunicorn = "^20.0.4"
supervisor = "^4.2.0"
psycopg2-binary = "^2.8.5"
djangorestframework-jwt = "^1.11.0"


[tool.poetry.dev-dependencies]
django-extensions = "^2.2.9"
notebook = "^6.0.3"
requests = "^2.24.0"
selenium = "^3.141.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
